Handle owner resolution failures
When running Hummingbird, backup is failing when resolving DLs using
either the AD lookup or the Get-DistributionList PowerShell cmdlet.
However, I am able to resolve members.

The code currently requires owner to be resolved before proceeding with
member lookup. I'm breaking this dependency so the user has the option
to still retrieve a backup of all members even if he/she cannot retrieve
the owner.